// src/app/services/openapi-client/model/productDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductDto { 
    id?: number;
    categoryId?: number | null;
    categoryName?: string | null;
    brandId?: number | null;
    brandName?: string | null;
    sku?: string | null;
    name?: string | null;
    slug?: string | null;
    shortDescription?: string | null;
    basePrice?: number;
    compareAtPrice?: number | null;
    isFeatured?: boolean;
    isBestseller?: boolean;
    isNewArrival?: boolean;
    isActive?: boolean;
    primaryImage?: string | null;
    averageRating?: number | null;
    reviewCount?: number;
    createdAt?: string | null;
}